+++ apps/www/src/modules/home/page/index.tsx
@@ -10,6 +10,7 @@ import BannerMarquee from "@/modules/common/components/banner-marquee";
 import Experience from "@/modules/common/components/experience";
 import FooterSetup from "@/modules/common/utilities/footer-setup";
 import { getHomePageData } from "../actions/home-page";
+import KeyFactsSection from "../widgets/key-facts";
 
 const HomePage = async () => {
   const data = await getHomePageData();
@@ -53,7 +54,23 @@ const HomePage = async () => {
         }
       })}
 
-      {/* <FeaturedServices
+      {/* <AboutUs
+        id={1}
+        __component="home.about-us"
+        with_container={true}
+        heading={{
+          id: 1,
+          title: "About Us",
+          sub_title: "We are a creative agency",
+        }}
+        content="Deft Dodo is a design & develop agency of the future with 120+ successful projects in their portfolio."
+      /> */}
+      {/* <Separator className="my-28" /> */}
+      <KeyFactsSection container />
+      <Separator className="mt-28 mb-10" />
+      {/* <Projects container /> */}
+      <Separator className="mt-28 mb-2" container />
+      {/*<FeaturedServices
         container
         heading="Know our services"
         subHeading="Featured Services"